unify tests to one file
